Rework confget()
Following an E-Mail exchange, I've written and tested this in various
ways, on my end, without any issues.

The behavior of the function should be the same as the previous code.
If it's expected to have thousands of lines of keys and values to parse
in the configuration file, then I suggest _not_ merging this commit.