Remove redundant quotes
Not a huge deal, I know, and I used to do the same thing myself for a
very long time, out of habit and without realising it wasn't needed, -
but none-the-less, the more you know!

With shell variable assignment and the WORD in a `case` statement, -
quotes aren't needed to protect them, as word-splitting doesn't occur.

For example:

```bash
NewVar=$OtherVar

NewVar="$OtherVar"
```

And that applies to most if not all standard expansions, like command
substitution, arithmetic expansion, brace expansion, etc.

However...

```bash
NewVar='This is a string.'

NewVar="This is $Num string."

NewVar=This is a string.

NewVar=This\ is\ a\ string.
```

So in _those_ sort of situations, the usual applies.